+++ src/Piece/PieceForm/PieceFields/PieceFields.js
@@ -33,6 +33,7 @@ export const PieceFields = ({
   instanceId,
   locationIds,
   locations,
+  isLocationsLoading,
   pieceFormatOptions,
   poLine,
   setLocationValue,
@@ -298,6 +299,7 @@ export const PieceFields = ({
             disabled={isLocationSelectionDisabled}
             isNonInteractive={isLocationSelectionDisabled}
             required={isLocationRequired}
+            isLoading={isLocationsLoading}
           />
         </Col>
       </Row>
@@ -308,12 +310,13 @@ export const PieceFields = ({
 PieceFields.propTypes = {
   createInventoryValues: PropTypes.object.isRequired,
   instanceId: PropTypes.string,
-  locationIds: PropTypes.arrayOf(PropTypes.string).isRequired,
-  locations: PropTypes.arrayOf(PropTypes.object),
   pieceFormatOptions: PropTypes.arrayOf(PropTypes.shape({
     label: PropTypes.string,
     value: PropTypes.string,
   })),
+  isLocationsLoading: PropTypes.bool,
+  locationIds: PropTypes.arrayOf(PropTypes.string),
+  locations: PropTypes.arrayOf(PropTypes.object),
   poLine: PropTypes.object.isRequired,
   setLocationValue: PropTypes.func.isRequired,
   onChangeDisplayOnHolding: PropTypes.func.isRequired,