+++ src/Piece/PieceForm/PieceForm.test.js
@@ -1,10 +1,8 @@
 import moment from 'moment';
-import { MemoryRouter } from 'react-router-dom';
 
 import user from '@folio/jest-config-stripes/testing-library/user-event';
 import {
   act,
-  render,
   screen,
   waitFor,
 } from '@folio/jest-config-stripes/testing-library/react';
@@ -15,6 +13,7 @@ import {
   PIECE_STATUS,
 } from '@folio/stripes-acq-components';
 
+import { renderWithRouter } from '../../../test/jest/helpers';
 import { usePieceStatusChangeLog } from '../hooks';
 import PieceForm from './PieceForm';
 
@@ -83,12 +82,11 @@ const logs = [
 const DATE_FORMAT = 'MM/DD/YYYY';
 const today = moment();
 
-const renderPieceForm = (props = {}) => render(
+const renderPieceForm = (props = {}) => renderWithRouter(
   <PieceForm
     {...defaultProps}
     {...props}
   />,
-  { wrapper: MemoryRouter },
 );
 
 const findButton = (name) => screen.findByRole('button', { name });