+++ src/Piece/PieceForm/PieceFormContainer.test.js
@@ -1,10 +1,5 @@
-import { MemoryRouter } from 'react-router-dom';
-
 import user from '@folio/jest-config-stripes/testing-library/user-event';
-import {
-  render,
-  screen,
-} from '@folio/jest-config-stripes/testing-library/react';
+import { screen } from '@folio/jest-config-stripes/testing-library/react';
 import {
   ORDER_FORMATS,
   ORDER_STATUSES,
@@ -14,6 +9,7 @@ import {
   useOrderLine,
 } from '@folio/stripes-acq-components';
 
+import { renderWithRouter } from '../../../test/jest/helpers';
 import {
   useHoldingItems,
   useOrder,
@@ -100,12 +96,11 @@ const defaultProps = {
   paneTitle: 'Piece form',
 };
 
-const renderPieceFormContainer = (props = {}) => render(
+const renderPieceFormContainer = (props = {}) => renderWithRouter(
   <PieceFormContainer
     {...defaultProps}
     {...props}
   />,
-  { wrapper: MemoryRouter },
 );
 
 describe('PieceFormContainer', () => {