+++ src/TitleForm/TitleForm.test.js
@@ -1,16 +1,6 @@
-import {
-  QueryClient,
-  QueryClientProvider,
-} from 'react-query';
-import {
-  MemoryRouter,
-  useHistory,
-} from 'react-router-dom';
+import { useHistory } from 'react-router-dom';
 
-import {
-  render,
-  screen,
-} from '@folio/jest-config-stripes/testing-library/react';
+import { screen } from '@folio/jest-config-stripes/testing-library/react';
 import user from '@folio/jest-config-stripes/testing-library/user-event';
 import {
   HasCommand,
@@ -19,6 +9,7 @@ import {
 } from '@folio/stripes/components';
 import { useOkapiKy } from '@folio/stripes/core';
 
+import { renderWithRouter } from '../../test/jest/helpers';
 import TitleForm from './TitleForm';
 
 jest.mock('@folio/stripes-components/lib/Commander', () => ({
@@ -59,20 +50,10 @@ const kyMock = {
   })),
 };
 
-const queryClient = new QueryClient();
-const wrapper = ({ children }) => (
-  <MemoryRouter>
-    <QueryClientProvider client={queryClient}>
-      {children}
-    </QueryClientProvider>
-  </MemoryRouter>
-);
-
-const renderTitleForm = (props = defaultProps) => render(
+const renderTitleForm = (props = defaultProps) => renderWithRouter(
   <TitleForm
     {...props}
   />,
-  { wrapper },
 );
 
 describe('TitleForm', () => {