+++ src/TitleReceive/TitleReceiveContainer.js
@@ -1,3 +1,4 @@
+import uniq from 'lodash/uniq';
 import {
   useCallback,
   useMemo,
@@ -12,11 +13,11 @@ import {
 } from '@folio/stripes/components';
 import {
   PIECE_FORMAT,
-  useLocationsQuery,
   useShowCallout,
 } from '@folio/stripes-acq-components';
 
 import {
+  useHoldingsAndLocations,
   useReceive,
   useTitleHydratedPieces,
 } from '../common/hooks';
@@ -37,7 +38,6 @@ function TitleReceiveContainer({ history, location, match }) {
   const showCallout = useShowCallout();
   const {
     crossTenant,
-    isCentralOrderingEnabled,
     isCentralRouting,
     targetTenantId,
   } = useReceivingSearchContext();
@@ -45,7 +45,7 @@ function TitleReceiveContainer({ history, location, match }) {
   const titleId = match.params.id;
 
   const {
-    pieces,
+    pieces = [],
     title,
     orderLine: poLine,
     isLoading: isPiecesLoading,
@@ -59,10 +59,22 @@ function TitleReceiveContainer({ history, location, match }) {
 
   const { receive } = useReceive();
 
-  const {
-    isLoading: isLocationsLoading,
-    locations,
-  } = useLocationsQuery({ consortium: isCentralOrderingEnabled });
+  const receivingTenantIds = useMemo(() => {
+    if (pieces?.length) {
+      return uniq([
+        ...pieces.map(({ receivingTenantId }) => receivingTenantId),
+        targetTenantId,
+      ].filter(Boolean));
+    }
+
+    return [];
+  }, [pieces, targetTenantId]);
+
+  const { locations, isFetching } = useHoldingsAndLocations({
+    instanceId,
+    receivingTenantIds,
+    tenantId: targetTenantId,
+  });
 
   const onCancel = useCallback(
     () => {
@@ -122,7 +134,7 @@ function TitleReceiveContainer({ history, location, match }) {
     [poLine],
   );
 
-  const isLoading = isPiecesLoading || isLocationsLoading;
+  const isLoading = isPiecesLoading || isFetching;
 
   if (isLoading) {
     return (