+++ src/common/hooks/useHoldingsAndLocations/useHoldingsAndLocations.js
@@ -1,23 +1,30 @@
 import { useQuery } from 'react-query';
 
-import {
-  batchRequest,
-  HOLDINGS_API,
-  LIMIT_MAX,
-  LOCATIONS_API,
-} from '@folio/stripes-acq-components';
+import { LIMIT_MAX } from '@folio/stripes-acq-components';
 import {
   useNamespace,
   useOkapiKy,
 } from '@folio/stripes/core';
 
+import {
+  getHoldingLocations,
+  getHoldingLocationsByTenants,
+} from '../../utils/getHoldingLocations';
+
 const DEFAULT_DATA = [];
 
-export const useHoldingsAndLocations = ({ instanceId, tenantId, options = {} }) => {
-  const {
-    enabled = true,
-    ...queryOptions
-  } = options;
+export const useHoldingsAndLocations = ({
+  instanceId,
+  options = {},
+  tenantId,
+  /*
+   `receivingTenantIds` is a unique list of tenantIds from the pieces list.
+    The purpose is that we need to be able to fetch locations from other
+    tenants so that we can display all the locations on the full-screen page
+  */
+  receivingTenantIds = DEFAULT_DATA,
+}) => {
+  const { enabled = true, ...queryOptions } = options;
 
   const ky = useOkapiKy({ tenant: tenantId });
   const [namespace] = useNamespace({ key: 'holdings-and-location' });
@@ -31,24 +38,11 @@ export const useHoldingsAndLocations = ({ instanceId, tenantId, options = {} }) 
     isFetching,
     isLoading,
   } = useQuery({
-    queryKey: [namespace, tenantId, instanceId],
-    queryFn: async ({ signal }) => {
-      const holdings = await ky
-        .get(HOLDINGS_API, { searchParams, signal })
-        .json()
-        .then(response => response.holdingsRecords);
-
-      const locationIds = holdings?.map(({ permanentLocationId }) => permanentLocationId) || DEFAULT_DATA;
-
-      const locations = await batchRequest(
-        ({ params }) => ky.get(LOCATIONS_API, { searchParams: params, signal }).json(),
-        locationIds,
-      ).then(responses => responses.flatMap(response => response.locations));
-
-      return {
-        holdings,
-        locations,
-      };
+    queryKey: [namespace, tenantId, instanceId, ...receivingTenantIds],
+    queryFn: ({ signal }) => {
+      return receivingTenantIds.length
+        ? getHoldingLocationsByTenants({ ky, instanceId, receivingTenantIds })
+        : getHoldingLocations({ ky, searchParams, signal });
     },
     enabled: enabled && Boolean(instanceId),
     ...queryOptions,
@@ -57,6 +51,7 @@ export const useHoldingsAndLocations = ({ instanceId, tenantId, options = {} }) 
   return ({
     holdings: data?.holdings || DEFAULT_DATA,
     locations: data?.locations || DEFAULT_DATA,
+    locationIds: data?.locationIds || DEFAULT_DATA,
     isFetching,
     isLoading,
   });