+++ src/common/hooks/useHoldingsAndLocations/useHoldingsAndLocations.test.js
@@ -7,9 +7,16 @@ import {
   renderHook,
   waitFor,
 } from '@folio/jest-config-stripes/testing-library/react';
+import { HOLDINGS_API } from '@folio/stripes-acq-components';
 import { useOkapiKy } from '@folio/stripes/core';
 
+import { extendKyWithTenant } from '../../utils';
 import { useHoldingsAndLocations } from './useHoldingsAndLocations';
+
+jest.mock('../../utils', () => ({
+  ...jest.requireActual('../../utils'),
+  extendKyWithTenant: jest.fn().mockReturnValue({ extend: jest.fn() }),
+}));
 
 const queryClient = new QueryClient();
 const wrapper = ({ children }) => (
@@ -45,16 +52,41 @@ const getMock = jest.fn()
 
 describe('useHoldingsAndLocations', () => {
   beforeEach(() => {
-    useOkapiKy.mockClear().mockReturnValue({
-      get: getMock,
-    });
+    useOkapiKy
+      .mockClear()
+      .mockReturnValue({
+        get: getMock,
+        extend: jest.fn(() => ({
+          get: jest.fn((path) => {
+            if (path === HOLDINGS_API) {
+              return {
+                json: jest.fn().mockResolvedValue({ holdingsRecords }),
+              };
+            }
+
+            return ({
+              json: jest.fn().mockResolvedValue({ locations }),
+            });
+          }),
+        })),
+      });
+    extendKyWithTenant.mockClear().mockReturnValue({ extend: jest.fn() });
   });
 
-  it('should fetch pieces requests', async () => {
+  it('should fetch holding locations', async () => {
     const { result } = renderHook(() => useHoldingsAndLocations({ instanceId: '1', tenantId: '2' }), { wrapper });
 
     await waitFor(() => expect(result.current.isLoading).toBeFalsy());
 
     expect(result.current.locations).toEqual(locations);
+  });
+
+  it('should fetch holding locations with different tenants', async () => {
+    const receivingTenantIds = ['1', '2'];
+    const { result } = renderHook(() => useHoldingsAndLocations({ instanceId: '1', receivingTenantIds, tenantId: '2' }), { wrapper });
+
+    await waitFor(() => expect(result.current.isLoading).toBeFalsy());
+
+    expect(result.current.locations).toHaveLength(2);
   });
 });